// frontend/src/api/model/trustScoreDTO.ts
/**
 * PropTech API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TrustScoreDTOFactorsInner } from './trustScoreDTOFactorsInner';


export interface TrustScoreDTO { 
    totalScore?: number;
    level?: TrustScoreDTO.LevelEnum;
    factors?: Array<TrustScoreDTOFactorsInner>;
}
export namespace TrustScoreDTO {
    export const LevelEnum = {
        Bronze: 'BRONZE',
        Silver: 'SILVER',
        Gold: 'GOLD',
        Platinum: 'PLATINUM'
    } as const;
    export type LevelEnum = typeof LevelEnum[keyof typeof LevelEnum];
}