// frontend/src/api/model/reputationScoreDTO.ts
/**
 * PropTech API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReviewDTO } from './reviewDTO';


export interface ReputationScoreDTO { 
    score?: number;
    level?: ReputationScoreDTO.LevelEnum;
    reviewCount?: number;
    contractReviewCount?: number;
    reviews?: Array<ReviewDTO>;
}
export namespace ReputationScoreDTO {
    export const LevelEnum = {
        SinHistorial: 'SIN_HISTORIAL',
        Bajo: 'BAJO',
        Medio: 'MEDIO',
        Bueno: 'BUENO',
        Excelente: 'EXCELENTE'
    } as const;
    export type LevelEnum = typeof LevelEnum[keyof typeof LevelEnum];
}