// frontend/src/api/model/solvencyResultDTO.ts
/**
 * PropTech API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SolvencyCheckDTO } from './solvencyCheckDTO';


export interface SolvencyResultDTO { 
    score?: number;
    level?: SolvencyResultDTO.LevelEnum;
    checks?: Array<SolvencyCheckDTO>;
    verifiedAt?: string;
    contractType?: SolvencyResultDTO.ContractTypeEnum | null;
    documentosCompletos?: boolean;
}
export namespace SolvencyResultDTO {
    export const LevelEnum = {
        Bronze: 'BRONZE',
        Silver: 'SILVER',
        Gold: 'GOLD',
        Platinum: 'PLATINUM'
    } as const;
    export type LevelEnum = typeof LevelEnum[keyof typeof LevelEnum];
    export const ContractTypeEnum = {
        Indefinido: 'INDEFINIDO',
        Temporal: 'TEMPORAL',
        Obra: 'OBRA'
    } as const;
    export type ContractTypeEnum = typeof ContractTypeEnum[keyof typeof ContractTypeEnum];
}